Make possible to specify the element to add the strength bar after
Before, the strength bar was always added after the element that
"strengthify" was called on. Now it is possible to specify the element
that the strength bar will be added after; if the "$addAfter" parameter
is not given then it is added after the element that "strengthify" was
called on just like before.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>