Add an option to disable the async zxcvbn.js loading
User can pass a `zxcvbn:null` option:

    $('#password-field').strengthify({zxcvbn: null})

In this case, assume zxcvbn.js is already loaded in the page
and ready to use.

Motivation: in my use case, zxcvbn.js is already included
in the JS bundle sent to the browser. It does not make sense
to load it again.